reorganized file structure, added delete account support
