--- frontend/src/components/home/editor/Link.tsx
@@ -1,7 +1,0 @@
-export default function Link({ element, attributes, children }: any) {
-	return (
-		<a href={element.url} {...attributes} className={"link"}>
-			{children}
-		</a>
-	);
-}
