+++ frontend/src/components/skeleton/SidebarNotelistSkeleton.tsx
@@ -1,0 +1,33 @@
+export default function SideBarNotelistSkeleton() {
+	return (
+		<div className="px-4 my-4 border-b border-gray-300 shadow-smb animate-pulse">
+			<div className="flex gap-4 place-items-center my-3">
+				<div className="flex flex-1 gap-3 flex-col">
+					<div className="h-4 bg-gray-200 rounded"></div>
+					<div className="h-4 bg-gray-200 rounded w-1/2"></div>
+				</div>
+				<div>
+					<div className="size-8 rounded-full bg-gray-200"></div>
+				</div>
+			</div>
+			<div className="flex gap-4 place-items-center my-8">
+				<div className="flex flex-1 gap-3 flex-col">
+					<div className="h-4 bg-gray-200 rounded"></div>
+					<div className="h-4 bg-gray-200 rounded w-1/2"></div>
+				</div>
+				<div>
+					<div className="size-8 rounded-full bg-gray-200"></div>
+				</div>
+			</div>
+			<div className="flex gap-4 place-items-center my-8">
+				<div className="flex flex-1 gap-3 flex-col">
+					<div className="h-4 bg-gray-200 rounded"></div>
+					<div className="h-4 bg-gray-200 rounded w-1/2"></div>
+				</div>
+				<div>
+					<div className="size-8 rounded-full bg-gray-200"></div>
+				</div>
+			</div>
+		</div>
+	);
+}
