+++ frontend/src/features/home/sidebar/CreateNote.tsx
