+++ frontend/src/features/home/sidebar/DeleteNoteDialog.tsx
